Refine processors
