+++ src/Microsoft.Health.Fhir.Liquid.Converter.UnitTests/ProcessorTests.cs
@@ -27,14 +27,14 @@ namespace Microsoft.Health.Fhir.Liquid.Converter.UnitTests
         {
             _hl7v2TestData = File.ReadAllText(Path.Join(TestConstants.SampleDataDirectory, "Hl7v2", "LRI_2.0-NG_CBC_Typ_Message.hl7"));
             _ccdaTestData = File.ReadAllText(Path.Join(TestConstants.SampleDataDirectory, "Ccda", "CCD.ccda"));
-            _jsonTestData = File.ReadAllText(Path.Join(TestConstants.SampleDataDirectory, "Json", "SimplePatient.json"));
+            _jsonTestData = File.ReadAllText(Path.Join(TestConstants.SampleDataDirectory, "Json", "ExamplePatient.json"));
         }
 
         public static IEnumerable<object[]> GetValidInputsWithTemplateDirectory()
         {
             yield return new object[] { new Hl7v2Processor(), new Hl7v2TemplateProvider(TestConstants.Hl7v2TemplateDirectory), _hl7v2TestData, "ORU_R01" };
             yield return new object[] { new CcdaProcessor(), new CcdaTemplateProvider(TestConstants.CcdaTemplateDirectory), _ccdaTestData, "CCD" };
-            yield return new object[] { new JsonProcessor(), new JsonTemplateProvider(TestConstants.JsonTemplateDirectory), _jsonTestData, "SimplePatient" };
+            yield return new object[] { new JsonProcessor(), new JsonTemplateProvider(TestConstants.JsonTemplateDirectory), _jsonTestData, "ExamplePatient" };
         }
 
         public static IEnumerable<object[]> GetValidInputsWithTemplateCollection()
